moved account handling to base activity since this is not file specific and thus should be available on (new) base level
